refactor(adapters): move spec from che168 to `autohome spec`
The new-car 参数配置 cache (cacheapigo.che168.com/CarProduct/GetParam) is
汽车之家-group data served unsigned and login-free — conceptually a sibling of
the existing no-login `autohome brand`/`score`, not the browser-gated che168
used-car adapter. Relocate it so che168 is purely used-car (browse/car) and
`autohome` owns every no-login 汽车之家 surface.

- new clis/autohome/spec.js (site: autohome); PARAM_API / SPEC_COLUMNS /
  normalizeSpecId / paramGetJson moved into autohome/utils.js
- che168/utils.js stripped to the used-car surface; header reworded
- param.json fixture + parseParams/normalizeSpecId tests moved che168 -> autohome
- docs: spec section moved to autohome.md; autohome's stale "no per-trim config"
  note corrected; cli-manifest.json regenerated
- autohome + che168 adapter tests green (19); live `autohome spec 39616` verified
