+++ clis/che168/utils.js
@@ -1,19 +1,14 @@
 /**
  * Shared helpers for the 汽车之家二手车 / 车168 (che168) adapter.
  *
- * che168 has two very different surfaces:
+ * che168's **used-car** listing/detail pages (`www.che168.com`) are fully gated
+ * behind a 瑞数 (Riversafe) JS challenge — a bare `fetch` only ever gets back a
+ * ~29KB challenge shell, never a listing. So the `browse` / `car` commands run
+ * in the logged-in browser (Strategy.DOM, `browser: true`), where the challenge
+ * clears naturally, and extract the rendered DOM.
  *
- *  1. The new-car **spec/config** data is served by an open, login-free cache
- *     API (`cacheapigo.che168.com/CarProduct/GetParam.ashx?specid=<id>`) that
- *     returns GBK-encoded JSON. The `spec` command reads this directly
- *     (Strategy.PUBLIC, no browser).
- *
- *  2. The **used-car** listing/detail pages (`www.che168.com` / `m.che168.com`)
- *     are fully gated behind a 瑞数 (Riversafe) JS challenge — a bare `fetch`
- *     only ever gets back a ~29KB challenge shell, never a listing. So the
- *     `browse` / `car` commands run in the logged-in browser (Strategy.DOM,
- *     `browser: true`), where the challenge clears naturally, and extract the
- *     rendered DOM.
+ * (New-car 参数配置 by specid is a separate, login-free cache surface and lives
+ * under `autohome spec` — same 汽车之家 group, no browser needed.)
  */
 
 import {
@@ -24,14 +19,7 @@ import {
 } from '@jackwener/opencli/errors';
 
 export const CHE168_WWW_BASE = 'https://www.che168.com';
-export const CHE168_M_BASE = 'https://m.che168.com';
-export const PARAM_API = 'https://cacheapigo.che168.com/CarProduct/GetParam.ashx';
 
-const UA =
-    'Mozilla/5.0 (iPhone; CPU iPhone OS 16_5 like Mac OS X) AppleWebKit/605.1.15 '
-    + '(KHTML, like Gecko) Version/16.5 Mobile/15E148 Safari/604.1';
-
-export const SPEC_COLUMNS = ['group', 'field', 'value'];
 export const BROWSE_COLUMNS = ['rank', 'info_id', 'title', 'price', 'reg_date', 'mileage', 'city', 'url'];
 export const CAR_COLUMNS = ['field', 'value'];
 
@@ -76,22 +64,6 @@ export function resolveCity(cityArg) {
 }
 
 /**
- * Normalize a che168/autohome spec id: a bare number, a `.../spec/<id>/` URL,
- * or a `?specid=<id>` query value.
- */
-export function normalizeSpecId(rawInput) {
-    const raw = String(rawInput ?? '').trim();
-    if (!raw) throw new ArgumentError('specid must be a non-empty value');
-    const m = raw.match(/specid[=/](\d+)/i) || raw.match(/spec[/_-](\d+)/i) || raw.match(/^(\d+)$/);
-    if (!m) {
-        throw new ArgumentError(
-            `'${rawInput}' does not look like a che168/autohome specid (a number, or a .../spec/<id>/ URL)`,
-        );
-    }
-    return m[1];
-}
-
-/**
  * Extract a che168 used-car listing id (infoid) for display: a bare number or
  * a `.../<infoid>.html` detail URL.
  */
@@ -125,39 +97,6 @@ export function resolveCarUrl(rawInput) {
         `'${rawInput}' is not a navigable che168 detail URL. Pass the full url from \`che168 browse\` `
         + '(e.g. https://www.che168.com/dealer/<dealer>/<infoid>.html) — a bare info id cannot reach the desktop detail page.',
     );
-}
-
-/** Fetch a GBK-encoded che168 cache-API endpoint and parse it as JSON. */
-export async function che168GetJson(url, contextHint) {
-    let resp;
-    try {
-        resp = await fetch(url, {
-            headers: {
-                'User-Agent': UA,
-                'Accept-Language': 'zh-CN,zh;q=0.9',
-                Referer: `${CHE168_M_BASE}/`,
-            },
-        });
-    } catch (err) {
-        throw new CommandExecutionError(`che168 ${contextHint} network error: ${err?.message || err}`);
-    }
-    if (!resp.ok) {
-        throw new CommandExecutionError(`che168 ${contextHint} HTTP ${resp.status}`);
-    }
-    const buf = await resp.arrayBuffer();
-    // che168 cache-API responses are GBK, not UTF-8.
-    const text = new TextDecoder('gbk').decode(buf);
-    if (/瑞数|reese84|安全验证|滑动验证/i.test(text) && !/returncode/i.test(text)) {
-        throw new AuthRequiredError(
-            'che168.com',
-            `che168 ${contextHint} hit an anti-bot challenge.`,
-        );
-    }
-    try {
-        return JSON.parse(text);
-    } catch {
-        throw new CommandExecutionError(`che168 ${contextHint} returned non-JSON (possibly anti-bot or schema change)`);
-    }
 }
 
 export { ArgumentError, AuthRequiredError, CommandExecutionError, EmptyResultError };
