+++ docs/adapters/browser/che168.md
@@ -1,28 +1,23 @@
 # 车168 / 汽车之家二手车 che168
 
-**Mode**: 🌐 Public (`spec`) · 🔐 Browser (`browse` / `car`) · **Domain**: `che168.com`
+**Mode**: 🔐 Browser · **Domain**: `che168.com`
 
-che168 (二手车之家, the used-car arm of the 汽车之家 / Autohome group) has two
-very different surfaces:
+che168 (二手车之家, the used-car arm of the 汽车之家 / Autohome group) serves its
+**used-car listing & detail pages** (`www.che168.com`) behind an anti-bot JS
+challenge; a plain HTTP client only ever gets the challenge shell. So `browse` /
+`car` run in the logged-in browser, where the challenge clears, and extract the
+rendered DOM.
 
-- **Spec/config data** is served by an open, login-free cache API
-  (`cacheapigo.che168.com/CarProduct/GetParam.ashx`, GBK JSON). `spec` reads it
-  directly — no browser.
-- **Used-car listing & detail pages** (`www.che168.com`) are gated behind an
-  anti-bot JS challenge; a plain HTTP client only ever gets the challenge shell.
-  So `browse` / `car` run in the logged-in browser, where the challenge clears,
-  and extract the rendered DOM.
+> New-car 参数配置 by specid is a separate, login-free surface and lives under
+> [`autohome spec`](./autohome.md) — same 汽车之家 group, no browser needed.
 
 ## Commands
 
-| Command | Mode | Description |
-|---------|------|-------------|
-| `opencli che168 spec <specid>` | 🌐 Public | Full car trim parameter sheet → 基本参数 / 车身 / 发动机 / 变速箱 / 底盘 / 制动 |
-| `opencli che168 browse [city]` | 🔐 Browser | Used cars for sale in a city → price / registration / mileage |
-| `opencli che168 car <url>` | 🔐 Browser | One listing's detail → price, down payment, registration, mileage, gearbox, transfers, location |
+| Command | Description |
+|---------|-------------|
+| `opencli che168 browse [city]` | Used cars for sale in a city → price / registration / mileage |
+| `opencli che168 car <url>` | One listing's detail → price, down payment, registration, mileage, gearbox, transfers, location |
 
-- `spec` takes a **specid** (the autohome/che168 trim id) — a bare number or any
-  `.../spec/<id>/` URL / `?specid=<id>` query.
 - `car` takes the listing **url** from `browse` (its `url` column). che168's
   desktop detail page needs the dealer-qualified path
   (`/dealer/<dealer>/<infoid>.html`); a bare info id can't reach it, so the full
@@ -31,11 +26,6 @@ very different surfaces:
 ## Usage Examples
 
 ```bash
-# Full spec/config sheet for a trim (open API, no browser)
-opencli che168 spec 39616
-opencli che168 spec https://www.autohome.com.cn/spec/39616/
-opencli che168 spec 39616 -f json
-
 # Browse used-car listings (defaults to Beijing)
 opencli che168 browse
 opencli che168 browse 上海 --limit 30
@@ -49,7 +39,6 @@ opencli che168 car https://www.che168.com/dealer/649418/58405825.html
 
 | Command | Columns |
 |---------|---------|
-| `spec` | `group, field, value` |
 | `browse` | `rank, info_id, title, price, reg_date, mileage, city, url` |
 | `car` | `field, value` (title, price, down_payment, reg_date, mileage, gearbox, fuel, transfers, location, publish_date, url) |
 
@@ -76,6 +65,5 @@ Pass a Chinese city name, a 2-letter code, or the che168 pinyin slug. Known:
 
 ## Prerequisites
 
-- `spec`: none — open API.
-- `browse` / `car`: a running browser bridge (`opencli doctor`). No che168 login
-  is required; the browser is used only to clear the anti-bot challenge.
+A running browser bridge (`opencli doctor`). No che168 login is required; the
+browser is used only to clear the anti-bot challenge.
